Reduced transform buffer size by 2x with no functionality loss

+++ DE2/src/CodingUnitStructure.c
@@ -76,7 +76,7 @@ void CodingUnitStructureConstructor(
 		&codingUnitStructure->transformBestBuffer,
 		codingUnitStructure->widthPicture,
 		codingUnitStructure->heightPicture,
-		sizeof(int));
+		sizeof(short));
 	BufferDescriptorConstructor(
 		&codingUnitStructure->reconBestBuffer,
 		codingUnitStructure->widthPicture,
